去掉crm后同步的代码
